+++ components/articles/styles/cardStyle.module.css
@@ -14,24 +14,15 @@
   overflow: hidden;
 }
 
-/* .articleContent {
-  display: flex;
-  flex-direction: column;
-  margin-top: 30px;
-  margin-left: 20px;
-} */
-/* .articleTitle {
-  font-size: 30px;
-  margin-bottom: 10px;
-  margin-left: 20px;
-} */
-/* 
-.articleDesc {
-  font-size: 15px;
-} */
+.articleCard:hover {
+  background-color: aliceblue;
+}
+
 .listContainer {
   display: flex;
   flex-direction: column;
+  padding-top: 100px;
+  padding-bottom: 100px;
 }
 
 .imageContainer {
@@ -45,6 +36,7 @@
   justify-content: center;
 
 }
+
 
 .buttonContainer {
   display: flex;
