+++ components/header/style/headerStyle.module.css
@@ -5,6 +5,8 @@
   background-color: black;
   position: fixed;
   width: 100%;
+
+  z-index: 10;
 }
 
 .items {
@@ -26,8 +28,11 @@
 }
 .logoContainer {
   margin-left: 15px;
-  color: aliceblue;
   align-self: center;
+}
+.logoContainer a {
+  text-decoration: none;
+  color: #fff;
 }
 
 .itemContainer{
