+++ styles/style.module.css
@@ -1,0 +1,4 @@
+.container {
+  padding-top: 100px;
+  padding-bottom: 50px;
+}
